Unified select project dialog into one file
